Update Snackbar not to use Wrapper Separate Midi file loader Apply some error to use snackbar
